ascksv/requirements.txt
asgiref==3.2.10
Django==3.0.8
pytz==2020.1
sqlparse==0.3.1
Pillow~=7.2.0
pytesseract~=0.3.4
djangorestframework~=3.11.0